+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # For next release
   * **Markus Freitag**
     * pkg/keystore: add convenience method for profile loading
+    * pkg/keepassxc: add convenience method for client initialization
 
 *Not released yet*